requirements.txt
python-dotenv==1.0.0
pythorhead==0.9.1
pytz==2023.3
pyyaml==6.0.0
requests==2.31.0
schedule==1.2.0